# swagger.json
{
	"components": {
...
		"schemas": {
			"UserI": {
				"properties": {
					"_id": {
						"type": "string",
						"description": "Mongo Document Id"
					},
					"username": {
						"type": "string",
						"description": "Username"
					},
					"email": {
						"type": "string",
						"description": "User email"
					},
					"userId": {
						"type": "string",
						"description": "Business User Id"
					},
					"password": {
						"type": "string",
						"description": "Password encrypted"
					},
					"firstName": {
						"type": "string",
						"description": "user first name"
					},
					"lastName": {
						"type": "string",
						"description": "User last name"
					},
					"active": {
						"type": "boolean",
						"description": "Is the user active?"
					},
					"externalUser": {
						"type": "boolean",
						"description": "External User. If true is external user that has participation in OpenChallenges"
					},
					"areaVisible": {
						"items": {
							"$ref": "#/components/schemas/AreaI"
						},
						"type": "array",
						"description": "Area that is visible for user"
					},
					"updated": {
						"type": "string",
						"format": "date-time",
						"description": "Last user update"
					}
				},
				"required": [
					"username",
					"email",
					"userId",
					"password",
					"firstName",
					"lastName",
					"active",
					"externalUser"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"AreaI": {
				"properties": {
					"name": {
						"type": "string"
					},
					"company": {
						"$ref": "#/components/schemas/CompanyI"
					}
				},
				"required": [
					"name",
					"company"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UserBody": {
				"description": "User interface used just for create a user",
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password",
					"email",
					"first_name",
					"last_name"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Login": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"SolutionI": {
				"properties": {
					"_id": {},
					"author": {
						"$ref": "#/components/schemas/UserI",
						"description": "Generator that create the solution. \nThis field exclusive with team configuration"
					},
					"team": {
						"$ref": "#/components/schemas/TeamI",
						"description": "Team that propose the solution. Just available for challenges\nwith TEAMWORK active.\nThis field exclusive with author and coauthor configuration"
					},
					"created": {
						"type": "string",
						"format": "date-time",
						"description": "Solution creation date"
					},
					"updated": {
						"type": "string",
						"format": "date-time",
						"description": "Solution update date"
					},
					"title": {
						"type": "string",
						"description": "Situation title"
					},
					"description": {
						"type": "string",
						"description": "Solution description"
					},
					"active": {
						"type": "boolean",
						"description": "Flag that indicate if a solution is active.\nWhen a challenge is delete, the flag is false"
					},
					"images": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "Solution images"
					},
					"WSALevel": {
						"type": "string",
						"enum": [
							"COMPANY",
							"AREA"
						],
						"description": "WorkSpaceAvailable.\nIf the situation is available for all company or just for some areas."
					},
					"areasAvailable": {
						"items": {
							"$ref": "#/components/schemas/AreaI"
						},
						"type": "array",
						"description": "Areas available. Is user if WSALevel = AREA because a situation isn't available for all Company. Just for some areas."
					},
					"departmentAffected": {
						"items": {
							"$ref": "#/components/schemas/AreaI"
						},
						"type": "array",
						"description": "Departments Affected"
					},
					"groupValidator": {
						"$ref": "#/components/schemas/GroupValidatorI",
						"description": "Group that do the analysis about the situation"
					},
					"status": {
						"type": "string",
						"description": "Challenge | Problem | Solution Status: @TODO define situation challenge"
					},
					"fileComplementary": {
						"type": "string",
						"description": "Complementary files to challenge, solution or problem"
					},
					"reactions": {
						"properties": {
							"comments": {
								"type": "number",
								"format": "double"
							},
							"confused": {
								"type": "number",
								"format": "double"
							},
							"likes": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"comments",
							"confused",
							"likes"
						],
						"type": "object",
						"description": "Reactions in park"
					},
					"solutionId": {
						"type": "string",
						"description": "Id that uniquely identifies a solution"
					},
					"challengeId": {
						"type": "string",
						"description": "Id that identify the challenge associated. If undefinedn then solution\nwas an initiative of a generator"
					},
					"challenge": {
						"$ref": "#/components/schemas/ChallengeI",
						"description": "Challenge Object. challengeId refer to challenge attribute. \nThe redundace is for performance lookup (for example Solutions with a particular challengeId)"
					},
					"canChooseScope": {
						"type": "boolean",
						"description": "If true, solution's generator can choose if the \nsolution is shared with the community"
					},
					"isPrivate": {
						"type": "boolean",
						"description": "Solution is shared with the community?"
					},
					"timeInPark": {
						"type": "number",
						"format": "double",
						"description": "TimePeriod that challenge is in a Park for discussion"
					}
				},
				"required": [
					"created",
					"title",
					"description",
					"active",
					"images",
					"WSALevel",
					"status",
					"fileComplementary",
					"solutionId",
					"canChooseScope",
					"isPrivate",
					"timeInPark"
				],
				"type": "object",
				"additionalProperties": true
			},
			"SolutionBody": {
				"properties": {
					"challenge_id": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"images": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"can_choose_scope": {
						"type": "boolean"
					},
					"is_private": {
						"type": "boolean"
					},
					"status": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"time_in_park": {
						"type": "number",
						"format": "double"
					},
					"baremo_validator": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"baremo_referrer": {
						"type": "string"
					},
					"file_name": {
						"type": "string"
					},
					"WSALevel": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"ChallengeResponse": {
				"properties": {
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"type": "string"
					},
					"updated": {
						"type": "string",
						"format": "date-time"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"images": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"WSALevel": {
						"type": "string"
					},
					"file_complementary": {
						"type": "string"
					},
					"reactions": {
						"properties": {
							"comments": {
								"type": "number",
								"format": "double"
							},
							"confused": {
								"type": "number",
								"format": "double"
							},
							"likes": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"comments",
							"confused",
							"likes"
						],
						"type": "object"
					},
					"is_strategic": {
						"type": "boolean"
					},
					"time_period": {
						"type": "number",
						"format": "double"
					},
					"participation_mode": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"created",
					"status",
					"updated",
					"title",
					"description",
					"images",
					"WSALevel",
					"file_complementary",
					"reactions",
					"is_strategic",
					"time_period",
					"participation_mode"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ChallengeBody": {
				"description": "Data that can be edited or inserted. Other are edited by \nanother endpoints",
				"properties": {
					"author": {
						"type": "string",
						"description": "author is required for a challenge"
					},
					"coauthor": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "Not used for challenge"
					},
					"team": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"images": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"WSALevel": {
						"type": "string",
						"enum": [
							"COMPANY",
							"AREA"
						]
					},
					"group_validator": {
						"type": "string"
					},
					"file_complementary": {
						"type": "string"
					},
					"is_strategic": {
						"type": "boolean"
					},
					"participation_mode": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"title",
					"description",
					"images",
					"WSALevel",
					"file_complementary",
					"author",
					"is_strategic",
					"participation_mode"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"CommitteI": {
				"properties": {
					"active": {
						"type": "boolean"
					},
					"leader": {
						"$ref": "#/components/schemas/UserI"
					},
					"general": {
						"items": {
							"$ref": "#/components/schemas/UserI"
						},
						"type": "array"
					}
				},
				"required": [
					"active"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"IntegrantI": {
				"properties": {
					"user": {
						"$ref": "#/components/schemas/UserI",
						"description": "Validator User"
					},
					"integrantId": {
						"type": "string",
						"description": "committe integrant Id"
					},
					"active": {
						"type": "boolean",
						"description": "is active?"
					},
					"created": {
						"type": "string",
						"format": "date-time",
						"description": "When was created?"
					},
					"lastChangePosition": {
						"type": "string",
						"format": "date-time",
						"description": "Date of the last change of position in the committe. \nFor example from LEADER to GENERAL or GENERAL to leader"
					},
					"finished": {
						"type": "string",
						"format": "date-time",
						"description": "Date end"
					},
					"groupValidator": {
						"$ref": "#/components/schemas/GroupValidatorI",
						"description": "committe integrant can be part of group validator\n. Just Committe integrants can be validators"
					},
					"role": {
						"type": "string",
						"description": "What is the role? Committe General or Committe Leader"
					}
				},
				"required": [
					"user",
					"integrantId",
					"active",
					"created",
					"lastChangePosition",
					"role"
				],
				"type": "object",
				"additionalProperties": true
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "ideation_backend",
		"version": "1.0.0",
		"description": "ideation backend",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Clara Idea"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/user/signup": {
			"post": {
				"operationId": "SignUp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserI"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserBody"
							}
						}
					}
				}
			}
		},
		"/user/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Login"
							}
						}
					}
				}
			}
		},
		"/user/{userId}": {
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/solution": {
			"post": {
				"operationId": "NewSolution",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SolutionI"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SolutionBody"
							}
						}
					}
				}
			}
		},
		"/solution/{solutionId}": {
			"patch": {
				"operationId": "UpdateSolutionPartially",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SolutionI"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "solutionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SolutionBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteSolution",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "solutionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"get": {
				"operationId": "GetSolution",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SolutionI"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "solutionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/challenge": {
			"post": {
				"operationId": "NewChallenge",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ChallengeResponse"
								}
							}
						}
					}
				},
				"description": "New Challenge method",
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Challenge definition according to ChallengeBody",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChallengeBody"
							}
						}
					}
				}
			}
		},
		"/challenge/{challengeId}/solution": {
			"post": {
				"operationId": "NewSolution",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SolutionI"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "challengeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChallengeBody"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ListSolutions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SolutionI"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "challengeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "init",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/challenge/{challengeId}": {
			"get": {
				"operationId": "GetChallenge",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ChallengeI"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "challengeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateChallengePartially",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ChallengeI"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "challengeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChallengeBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteChallenge",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "challengeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/committe/leader": {
			"post": {
				"operationId": "NewLeader",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommitteI"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "string"
							}
						}
					}
				}
			}
		},
...
		"/integrant/leader/{integrantId}": {
			"post": {
				"operationId": "NewLeader",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IntegrantI"
								}
							}
						}
					}
				},
				"description": "Endpoint for add or reeplace a leader of committe.\nThe new leader have to be a `general commite` user first",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "integrantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/integrant/general": {
			"post": {
				"operationId": "NewIntegrant",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IntegrantI"
								}
							}
						}
					}
				},
				"description": "Endpoint used for add a new integrant to committe. \nUser could be a new integrant or old integrant with active=false",
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "user that will be converted from generator to commite (GENERAL)",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "string"
							}
						}
					}
				}
			}
		},
...
		"/integrant/general/{integrantId}": {
			"delete": {
				"operationId": "DeleteLeader",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "integrantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}